rework Scheduled BATs regarding Location
Switching over to using the Windows Environment Variable: %computername% to get the name of the system. Settings.ini entries should match that same name. And then PRIVATE will have subfolders that are specific to each system (for backups, etc). And then that can be synced to cloud, Resilio, etc. for extra layer of backups.