adding INITOOL.EXE
This tool (https://github.com/dbohdan/initool) allows read/write of INI files from command line. Should help with implementing the rootFolder hard-coding removal across BAT files as well. The code in the 'testing-initool.cmd' is how you read the value and set it to a variable in BAT/CMD's