Update BPTV-CHEAT-SHEETS.ahk
CHANGE IN FUNCTIONALITY - I've changed the way that you get out of a Cheat Sheet. Previously the script waited until you released whichever hotkey you used to show a given sheet. That was getting annoying to hold down the key the whole time. Now I've changed it so that Escape will close the open cheat sheet. Beware that the cheatsheets are set to be 'alwaysontop'

I've also added a Tooltip that reminds you that Escape is how you close the sheets. And how to switch tabs for the KBF2 cheaetsheet (the app-specific image one).