Update BPTV-CHEAT-SHEETS.ahk
Updated 'firstShower' to have the ability to display a text file with a vertical scroll bar. This was needed because the PRIVATE\MASTER-SCAF-KEYS.ahk file grew longer than the height of the Cheat Sheet. (This file is not in the repo - I porbably need to add some example of the things I have outside the repo or gitignored because they contain private info).