Update PREMIERE-PRO-HOTKEYS.ahk
Some hotkeys pertaining to a specific project, but there's a new concept in here for editing clip/sequence names in the Project Panel by using RegExReplace to find&replace text. I think this could be expanded. However, 'InputBox' kept auto-invoking itself and sometimes would get stuck in a loop and would not close. Probably would have to be done as a GUI.