Began adding Resolve features
Resolve has a very complex UI. Several items in the UI are not directly addressable via keyboard shortcuts. You can't even assign some of them shortcut keys. To solve this, we use 1-shot AHK scripts that will use a pre-saved set of locations to make these items become addressable via hotkeys or StreamDeck buttons. This is the start of that effort.