small tweaks to the grabTC functions
This function was not working correctly when using CoorMode 'window' or 'client' - the only way I could get it to work consistently is to use 'screen'. Ultimately that's probably OK since we don't store the xPos & yPos to grab across systems. They are pretty much unique values to a specific system.